CreatureKilledGoal renamed to ConsumeCorpse. The old one got a new name CorpseConsumed. From now on can detect the kill even if the app is not active but increment the kill count.
