+++ Core/Goals/ConsumeCorpse.cs
@@ -1,34 +1,52 @@
 ﻿using System.Threading.Tasks;
-using Core.GOAP;
 using Microsoft.Extensions.Logging;
+using Core.GOAP;
 
 namespace Core.Goals
 {
-    class ConsumeCorpse : GoapGoal
+    public class ConsumeCorpse : GoapGoal
     {
-        public override float CostOfPerformingAction { get => 4.7f; }
-        public override bool Repeatable => false;
+        public override float CostOfPerformingAction { get => 4.1f; }
+
+        public override bool Repeatable { get; } = false;
 
         private readonly ILogger logger;
-        private readonly GoapAgentState goapAgentState;
+        private readonly ClassConfiguration classConfig;
 
-        public ConsumeCorpse(ILogger logger, GoapAgentState goapAgentSate)
+        public ConsumeCorpse(ILogger logger, ClassConfiguration classConfig)
         {
             this.logger = logger;
-            this.goapAgentState = goapAgentSate;
+            this.classConfig = classConfig;
 
-            AddPrecondition(GoapKey.incombat, false);
-            AddPrecondition(GoapKey.consumecorpse, true);
+            AddPrecondition(GoapKey.dangercombat, false);
 
-            AddEffect(GoapKey.consumecorpse, false);
+            AddPrecondition(GoapKey.producedcorpse, true);
+            AddPrecondition(GoapKey.consumecorpse, false);
+
+            AddEffect(GoapKey.producedcorpse, false);
+            AddEffect(GoapKey.consumecorpse, true);
+
+            if (classConfig.Loot)
+            {
+                AddEffect(GoapKey.shouldloot, true);
+
+                if (classConfig.Skin)
+                {
+                    AddEffect(GoapKey.shouldskin, true);
+                }
+            }
         }
 
         public override async Task PerformAction()
         {
-            goapAgentState.DecKillCount();
-            logger.LogInformation($"----- Consumed a corpse. Remaining: {goapAgentState.LastCombatKillCount}");
+            logger.LogWarning("----- Safe to consume a corpse.");
+            SendActionEvent(new ActionEventArgs(GoapKey.consumecorpse, true));
 
-            SendActionEvent(new ActionEventArgs(GoapKey.consumecorpse, false));
+            if (classConfig.Loot)
+            {
+                SendActionEvent(new ActionEventArgs(GoapKey.shouldloot, true));
+            }
+
             await Task.Delay(5);
         }
     }
